requirements.txt
python-telegram-bot[job-queue]
python-dotenv
openai
google-api-python-client
google-auth-oauthlib